Extract readable tile archive interface
